Cleaned up binding refactoring
* fixed up constructors
* fixed up unit tests
